backend/requirements.txt
# Backend Requirements for Car Loan Assistant

fastapi>=0.136.0
uvicorn>=0.44.0
python-multipart>=0.0.26
pydantic>=2.13.2
requests>=2.33.1
openai>=2.32.0
PyPDF2==3.0.1
pytesseract>=0.3.13
pdf2image>=1.17.0
Pillow>=12.2.0
pdfminer.six>=20260107
python-dotenv>=1.2.2
python-docx>=1.2.0
spacy>=3.8.7
langchain>=0.3.27
langchain-openai>=0.3.35
